requirements-dev.txt
pytest-qt
pytest-mock
importlib-metadata
sqlalchemy
pytest-cov
bcrypt